Completed pickPhysicalDevice()
Here is where Vulkan will pick the physical GPU that will be used for the program.

I added code in here to ensure that it chooses the discrete graphics card if the computer is plugged in. If not, it will chose the integrated one to save battery power.

In the future, more sophisticated checking will need to be added.