Added checks for GPU suitability
Added extension support checks to ensure that the physical device used by Vulkan is suitable given the required extensions.

This should later on be updated to include a score system with optional extensions as well. This is so that the game engine can pick the best possible device to use.