Set Up Skeleton For Model Classes
These model classes will be used to store for vertex data for each graphical object.

Every object/sprite that is drawn onto the screen in Vulkan, is drawn using both vertex and color data. This Vertex and color data is then fed into the graphics pipeline where it is rendered onto the screen.