fixed failed swap chain error
an error, that would not create the swap chain when ran on a device with an AMD GPU, would occur because the code did not chose an alpha mode based on the physical device. I fixed the error by allowing the appropriate alpha mode to be selected based on the device. I also included more print statements that would hopefully provide better help for any future swap chain issues